\name{DiscreteKL-package}
\alias{DiscreteKL-package}
\alias{DiscreteKL}
\docType{package}
\title{
  Kullback-Leibler-Based Discrete Failure Time Models
}
\description{
  Prediction models built based on a single data source may suffer from rare event rates, small sample sizes, and low signal-to-noise ratios. Incorporating published prediction models from large-scale studies is expected to improve the performance of prognosis prediction. Thus, we propose Kullback-Leibler-based (KL) discrete failure time models to integrate published prediction models (external models) with an individual-level time-to-event dataset (internal data) while accounting potential heterogeneity among different information sources.
}
\author{
Di Wang

Maintainer: Di Wang <dwwang@umich.edu>
}

